requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.2
google-auth==1.21.2
requests==2.24.0
google-cloud-vision
flask_login